requirements.txt
# --- Backend API ---
fastapi
uvicorn
python-multipart
python-dotenv

# --- Quantized LLM (Gemma-3) ---
transformers @ git+https://github.com/huggingface/transformers@v4.49.0-Gemma-3
torch
bitsandbytes
accelerate

# --- LangChain for RAG ---
langchain
langchain-community
langchain-huggingface
sentence-transformers
accelerate>=0.26.0
langchain-chroma>=0.1.2


# --- File Parsing & Loaders ---
pypdf
python-docx
docx2txt

# --- Frontend ---
streamlit
requests
